Moved shared modules to lib folder in examples
