+++ examples/lib/fb.h
