+++ examples/lib/interpolate.c
